# uapi-swagger-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "User API",
    "description": "API спецификация веб-приложения проекта User API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "PUBLIC: Пользователи"
        ],
        "summary": "Получение списка пользователей",
        "description": "Позволяет получить список пользователей. Есть возможность фильтрации по параметрам",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "firstName",
            "in": "query",
            "description": "Фильтр по имени",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastName",
            "in": "query",
            "description": "Фильтр по фамилии",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patronymic",
            "in": "query",
            "description": "Фильтр по отчеству",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "yearOfBirth",
            "in": "query",
            "description": "Фильтр по году рождения",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Количество элементов, которые нужно пропустить для формирования текущего набора",
            "required": true,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Количество элементов в наборе",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "400 BAD_REQUEST",
                  "reason": "Ошибка валидации данных из запроса",
                  "message": "getAllUserCommunity.from: Параметр запроса from, должен быть положительным числом или нулём.",
                  "timestamp": "2024-08-10 04:32:51"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "500 INTERNAL_SERVER_ERROR",
                  "reason": "Произошла непредвиденная ошибка",
                  "message": "Пожалуйста, обратитесь в службу технической поддержки",
                  "timestamp": "2024-08-10 01:33:59"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FullUserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PUBLIC: Пользователи"
        ],
        "summary": "Регистрация пользователя",
        "description": "Позволяет зарегистрировать пользователя",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "500 INTERNAL_SERVER_ERROR",
                  "reason": "Произошла непредвиденная ошибка",
                  "message": "Пожалуйста, обратитесь в службу технической поддержки",
                  "timestamp": "2024-08-10 01:33:59"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewUserResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "409 CONFLICT",
                  "reason": "Объект уже существует",
                  "message": "Пользователь с таким email уже существует. email = ivanov@mail.ru",
                  "timestamp": "2024-08-10 01:48:15"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "400 BAD_REQUEST",
                  "reason": "Ошибка валидации данных из запроса",
                  "message": "Некорректный адрес электронной почты: ivanov.ru",
                  "timestamp": "2024-08-10 01:32:27"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/profile/update": {
      "patch": {
        "tags": [
          "USER: Пользователи"
        ],
        "summary": "Обновление данных пользователя",
        "description": "Позволяет обновить данные пользователя",
        "operationId": "updateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "500 INTERNAL_SERVER_ERROR",
                  "reason": "Произошла непредвиденная ошибка",
                  "message": "Пожалуйста, обратитесь в службу технической поддержки",
                  "timestamp": "2024-08-10 01:33:59"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "403 FORBIDDEN",
                  "reason": "Ошибка валидации токена",
                  "message": "Срок действия токена истёк",
                  "timestamp": "2024-08-10 05:16:06"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullUserDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "404 NOT_FOUND",
                  "reason": "Объект не найден",
                  "message": "Пользователя с таким id не существует. userId = 56",
                  "timestamp": "2024-08-10 05:01:47"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "409 CONFLICT",
                  "reason": "Объект уже существует",
                  "message": "Пользователь с таким email уже существует. email = ivanov@mail.ru",
                  "timestamp": "2024-08-10 01:48:15"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "400 BAD_REQUEST",
                  "reason": "Ошибка валидации данных из запроса",
                  "message": "Некорректный адрес электронной почты: ivanov.ru",
                  "timestamp": "2024-08-10 01:32:27"
                }
              }
            }
          }
        }
      }
    },
    "/users/profile/{userId}": {
      "get": {
        "tags": [
          "PUBLIC: Пользователи"
        ],
        "summary": "Получение профиля пользователя",
        "description": "Позволяет получить данные одного из пользователей",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "500 INTERNAL_SERVER_ERROR",
                  "reason": "Произошла непредвиденная ошибка",
                  "message": "Пожалуйста, обратитесь в службу технической поддержки",
                  "timestamp": "2024-08-10 01:33:59"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "400 BAD_REQUEST",
                  "reason": "Ошибка валидации данных из запроса",
                  "message": "getOtherUserById.userId: Параметр запроса userId, должен быть положительным числом.",
                  "timestamp": "2024-08-10 04:56:50"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "404 NOT_FOUND",
                  "reason": "Объект не найден",
                  "message": "Пользователя с таким id не существует. userId = 56",
                  "timestamp": "2024-08-10 05:01:47"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{userId}": {
      "delete": {
        "tags": [
          "ADMIN: Пользователи"
        ],
        "summary": "Удаление пользователя",
        "description": "Позволяет удалить пользователя из БД",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "500 INTERNAL_SERVER_ERROR",
                  "reason": "Произошла непредвиденная ошибка",
                  "message": "Пожалуйста, обратитесь в службу технической поддержки",
                  "timestamp": "2024-08-10 01:33:59"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "403 FORBIDDEN",
                  "reason": "Ошибка валидации токена",
                  "message": "Срок действия токена истёк",
                  "timestamp": "2024-08-10 05:16:06"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "400 BAD_REQUEST",
                  "reason": "Ошибка валидации данных из запроса",
                  "message": "getOtherUserById.userId: Параметр запроса userId, должен быть положительным числом.",
                  "timestamp": "2024-08-10 04:56:50"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "404 NOT_FOUND",
                  "reason": "Объект не найден",
                  "message": "Пользователя с таким id не существует. userId = 56",
                  "timestamp": "2024-08-10 05:01:47"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "NewUserRequest": {
        "required": [
          "email",
          "firstName",
          "password"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван"
          },
          "email": {
            "maxLength": 254,
            "minLength": 6,
            "type": "string",
            "description": "Электронная почта",
            "example": "ivan@mail.ru"
          },
          "password": {
            "maxLength": 50,
            "minLength": 4,
            "type": "string",
            "description": "Пароль от аккаунта пользователя",
            "example": "g32jhg4jh23"
          }
        },
        "description": "Данные нового пользователя"
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Код ошибки и статус",
            "example": "404 NOT_FOUND"
          },
          "reason": {
            "type": "string",
            "description": "Причина ошибки",
            "example": "Объект не найден"
          },
          "message": {
            "type": "string",
            "description": "Подробное описание ошибки",
            "example": "Пользователя с таким id не существует. userId = 99"
          },
          "timestamp": {
            "type": "string",
            "description": "Дата и время, когда произошла ошибка",
            "example": "2024-07-27 06:27:23"
          }
        },
        "description": "Описание ошибки"
      },
      "NewUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "id пользователя",
            "format": "int64",
            "example": 7
          },
          "firstName": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван"
          },
          "email": {
            "type": "string",
            "description": "Электронная почта",
            "example": "ivan@mail.ru"
          },
          "registrationDate": {
            "type": "string",
            "description": "Дата регистрации пользователя",
            "example": "1995-03-27 22:13:36"
          }
        },
        "description": "Ответ на создание нового пользователя"
      },
...
      "UpdateUserRequest": {
        "required": [
          "email",
          "firstName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван"
          },
          "lastName": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string",
            "description": "Фамилия пользователя",
            "example": "Иванов"
          },
          "patronymic": {
            "maxLength": 50,
            "minLength": 2,
            "type": "string",
            "description": "Отчество пользователя",
            "example": "Иванович"
          },
          "birthday": {
            "type": "string",
            "description": "Дата рождения пользователя",
            "format": "date",
            "example": "1995-03-27"
          },
          "email": {
            "maxLength": 254,
            "minLength": 6,
            "type": "string",
            "description": "Электронная почта",
            "example": "ivan@mail.ru"
          },
          "phone": {
            "maxLength": 13,
            "minLength": 11,
            "type": "string",
            "description": "Номер телефона пользователя",
            "example": "88005555535"
          }
        },
        "description": "Обновление информации о пользователе"
      },
      "FullUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "id пользователя",
            "format": "int64",
            "example": 7
          },
          "firstName": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия пользователя",
            "example": "Иванов"
          },
          "patronymic": {
            "type": "string",
            "description": "Отчество пользователя",
            "example": "Иванович"
          },
          "birthday": {
            "type": "string",
            "description": "Дата рождения пользователя",
            "example": "1995.03.27"
          },
          "email": {
            "type": "string",
            "description": "Электронная почта",
            "example": "ivan@mail.ru"
          },
          "phone": {
            "type": "string",
            "description": "Номер телефона пользователя",
            "example": "8 800 555 55 35"
          }
        },
        "description": "Вывод всей информации о пользователе"
      },
...
    }
  }
}